Add option to test with FFTW backend
