+++ .github/workflows/CI.yml
@@ -22,6 +22,12 @@ jobs:
           - windows-latest
         arch:
           - x64
+        group:
+          - TestPlans
+          - FFTW 
+        exclude:
+          - version: '1.0'
+            group: FFTW
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
@@ -40,7 +46,10 @@ jobs:
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
+        env:
+          GROUP: ${{ matrix.group }}
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
         with:
           file: lcov.info
+          flag-name: group-${{ matrix.group }} # unique name for coverage report of each group
